Update unit tests

+++ bitwarden_license/src/Commercial.Core/SecretsManager/Commands/AccessPolicies/DeleteAccessPolicyCommand.cs
@@ -1,4 +1,5 @@
-﻿using Bit.Core.SecretsManager.Commands.AccessPolicies.Interfaces;
+﻿using Bit.Core.Exceptions;
+using Bit.Core.SecretsManager.Commands.AccessPolicies.Interfaces;
 using Bit.Core.SecretsManager.Repositories;
 
 namespace Bit.Commercial.Core.SecretsManager.Commands.AccessPolicies;
@@ -16,6 +17,12 @@ public class DeleteAccessPolicyCommand : IDeleteAccessPolicyCommand
 
     public async Task DeleteAsync(Guid id)
     {
+        var accessPolicy = await _accessPolicyRepository.GetByIdAsync(id);
+        if (accessPolicy == null)
+        {
+            throw new NotFoundException();
+        }
+
         await _accessPolicyRepository.DeleteAsync(id);
     }
 }
